backend/requirements.txt
# Django Core
Django==5.1.3
djangorestframework==3.15.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.63.2
dj-rest-auth
python-decouple==3.8
django-otp==1.3.0
django-two-factor-auth==1.16.0
pyotp==2.9.0
phonenumbers==8.13.34  # Required by django-two-factor-auth

# Email Sending
django-anymail==10.2  # Optional for advanced email services like SendGrid, Mailgun

# QR Code Generation for 2FA
qrcode==7.4.2
Pillow==10.3.0

# Async Task Management (Optional for Notifications, Background Tasks)
celery==5.4.0
redis==5.0.4
